refactor: extract normalizeUrl and urlToId to @usopc/shared (#200)
Deduplicate URL normalization and ID generation that was copied between
packages/core and packages/ingestion. Both packages now import from the
single shared implementation.
