+++ packages/shared/src/index.ts
@@ -89,3 +89,5 @@ export type {
   OrgType,
   SportOrganization,
 } from "./types/sport-org.js";
+
+export { normalizeUrl, urlToId } from "./url.js";
